Partial overhaul of exception handling in the REST client. Because 'errors' (as opposed to exceptional exceptions) need to be sent back with some useful info to the client (not just a stacktrace) we're now using Jerset exception mappers to handle them. This might soon be extended to Serialization/ Deserialization errors.
Additionally, added proper setup/teardown for individual REST tests.

+++ ehri-extension/src/test/java/eu/ehri/extension/test/AdminRestClientTest.java
@@ -6,8 +6,6 @@ import static org.junit.Assert.assertTrue;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -19,51 +17,57 @@ import eu.ehri.extension.AdminResource;
 import eu.ehri.project.models.base.AccessibleEntity;
 import eu.ehri.project.persistance.EntityBundle;
 
+/**
+ * Test admin REST functions.
+ * 
+ * @author michaelb
+ * 
+ */
 public class AdminRestClientTest extends BaseRestClientTest {
 
-    
-    
     @BeforeClass
     public static void setUpBeforeClass() throws Exception {
         initializeTestDb(AdminRestClientTest.class.getName());
     }
-    
+
     @Test
     public void testCreateDefaultUser() throws Exception {
         // Create
         WebResource resource = client.resource(getExtensionEntryPointUri()
                 + "/admin/createDefaultUserProfile");
-        ClientResponse response = resource
-                .accept(MediaType.APPLICATION_JSON)
-                .type(MediaType.APPLICATION_JSON)
-                .post(ClientResponse.class);
+        ClientResponse response = resource.accept(MediaType.APPLICATION_JSON)
+                .type(MediaType.APPLICATION_JSON).post(ClientResponse.class);
 
-        assertEquals(Response.Status.CREATED.getStatusCode(), response.getStatus());
+        assertEquals(Response.Status.CREATED.getStatusCode(),
+                response.getStatus());
         String json = response.getEntity(String.class);
         EntityBundle<VertexFrame> bundle = converter.jsonToBundle(json);
-        String ident = (String) bundle.getData().get(AccessibleEntity.IDENTIFIER_KEY);
+        String ident = (String) bundle.getData().get(
+                AccessibleEntity.IDENTIFIER_KEY);
         assertTrue(ident != null);
         assertTrue(ident.startsWith(AdminResource.DEFAULT_USER_ID_PREFIX));
-        
-        // Create another user and ensure their idents are different and incremental
+
+        // Create another user and ensure their idents are different and
+        // incremental
         WebResource resource2 = client.resource(getExtensionEntryPointUri()
                 + "/admin/createDefaultUserProfile");
-        ClientResponse response2 = resource2
-                .accept(MediaType.APPLICATION_JSON)
-                .type(MediaType.APPLICATION_JSON)
-                .post(ClientResponse.class);
+        ClientResponse response2 = resource2.accept(MediaType.APPLICATION_JSON)
+                .type(MediaType.APPLICATION_JSON).post(ClientResponse.class);
 
-        assertEquals(Response.Status.CREATED.getStatusCode(), response2.getStatus());
+        assertEquals(Response.Status.CREATED.getStatusCode(),
+                response2.getStatus());
         String json2 = response2.getEntity(String.class);
         EntityBundle<VertexFrame> bundle2 = converter.jsonToBundle(json2);
-        String ident2 = (String) bundle2.getData().get(AccessibleEntity.IDENTIFIER_KEY);
+        String ident2 = (String) bundle2.getData().get(
+                AccessibleEntity.IDENTIFIER_KEY);
         assertEquals(parseUserId(ident) + 1L, parseUserId(ident2));
         assertTrue(ident.startsWith(AdminResource.DEFAULT_USER_ID_PREFIX));
-        
+
     }
-    
+
     // Helpers
     private long parseUserId(String ident) {
-        return (long)Long.parseLong(ident.replace(AdminResource.DEFAULT_USER_ID_PREFIX, ""));
+        return Long.parseLong(ident.replace(
+                AdminResource.DEFAULT_USER_ID_PREFIX, ""));
     }
 }
